+++ app/mailers/user_mailer.rb
@@ -2,7 +2,7 @@ class UserMailer < ApplicationMailer
 
   def contact(name, email, message)
     subject = "#{name} - Website Message"
-    @email_to = 'reels@conductorproductions.com'
+    @email_to = ENV['contact_email']
     @message = message
     mail(to: @email_to, subject: subject, reply_to: email)
   end