+++ db/migrate/20150803235703_add_archived_to_projects.rb
@@ -1,0 +1,5 @@
+class AddArchivedToProjects < ActiveRecord::Migration
+  def change
+    add_column :projects, :archived, :boolean, null: false, default: false
+  end
+end